+++ Styx_Station/Assets/03. Scripts/States/MonsterTypeBase.cs
@@ -1,10 +1,11 @@
+using System.Numerics;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "MonsterTypeBase.asset", menuName = "Monster/MonsterBase")]
 public class MonsterTypeBase : ScriptableObject
 {
-    public float maxHealth;
-    public float damage;
+    public BigInteger maxHealth;
+    public BigInteger damage;
     public AttackType attackType;
     public GameObject prefab;
     public float speed;
