+++ Styx_Station/Assets/03. Scripts/Struct/Attack.cs
@@ -1,12 +1,13 @@
 using System.Collections;
 using System.Collections.Generic;
+using System.Numerics;
 using UnityEngine;
 
 public struct Attack 
 {
-    public int Damage { get; private set; }
+    public BigInteger Damage { get; private set; }
     public bool IsCritical { get; private set; }    
-    public Attack(int damage, bool critical)
+    public Attack(BigInteger damage, bool critical)
     {
         Damage = damage;
         IsCritical = critical;
